backend/blogweb/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.38.0
cryptography==42.0.2
Deprecated==1.2.14
Django==5.0.1
django-admin-interface==0.29.1
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-jet==1.0.8
django-js-asset==2.2.0
django-material==1.12.0
django-oauth-toolkit==2.3.0
django-sendgrid==1.0.1
django-sendgrid-v5==1.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
jwcrypto==1.5.1
oauthlib==3.2.2
packaging==23.2
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.21
PyJWT==2.9.0
PyMySQL==1.1.0
python-http-client==3.3.7
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sendgrid==6.11.0
setuptools==75.3.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
text-unidecode==1.3
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
wrapt==1.16.0
